applications/api/src/main/java/com/rps/stats/StatsController.java
package com.rps.stats;

import lombok.RequiredArgsConstructor;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

import java.util.List;

import static org.springframework.http.MediaType.APPLICATION_JSON_VALUE;

@RestController
@RequestMapping("/stats")
@RequiredArgsConstructor
public class StatsController {

    private final StatsService statsService;

    @GetMapping(produces = APPLICATION_JSON_VALUE)
    public List<PlayerStats> getAll() {
        return statsService.getAll();
    }

    @GetMapping(path = "/{id}", produces = APPLICATION_JSON_VALUE)
    public PlayerStats getById(@PathVariable Integer id) {
        return statsService.getById(id);
    }
}
